+++ packages/forma-36-react-components/src/components/Form/FieldGroup/FieldGroup.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
+import { storiesOf, StoryDecorator } from '@storybook/react';
 import { boolean } from '@storybook/addon-knobs';
 import { StateDecorator, Store } from '@sambego/storybook-state';
 import { host } from 'storybook-host';
@@ -20,7 +20,7 @@ storiesOf('Components|Form/FieldGroup', module)
       width: 400,
     }),
   )
-  .addDecorator(StateDecorator(store))
+  .addDecorator(StateDecorator(store) as StoryDecorator)
   .add(
     'default',
     withInfo()(() => (
@@ -29,14 +29,18 @@ storiesOf('Components|Form/FieldGroup', module)
           labelText="I agree"
           value="yes"
           helpText="Click if you agree"
-          onChange={e => store.set({ agreeTerms: e.target.value })}
+          onChange={e =>
+            store.set({ agreeTerms: (e.target as HTMLInputElement).value })
+          }
           checked={store.state.agreeTerms === 'yes'}
           id="termsCheckboxYes"
         />
         <CheckboxField
           labelText="I don't agree"
           value="no"
-          onChange={e => store.set({ agreeTerms: e.target.value })}
+          onChange={e =>
+            store.set({ agreeTerms: (e.target as HTMLInputElement).value })
+          }
           checked={store.state.agreeTerms === 'no'}
           helpText="Click if you don't agree"
           id="termsCheckboxNo"
