+++ packages/forma-36-react-components/src/components/Form/Form/Form.stories.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
+import { storiesOf, StoryDecorator } from '@storybook/react';
 import { host } from 'storybook-host';
 import { withInfo } from '@storybook/addon-info';
 import { StateDecorator, Store } from '@sambego/storybook-state';
 import { action } from '@storybook/addon-actions';
 import { boolean, select } from '@storybook/addon-knobs';
 
-import Form from '.';
+import Form from './Form';
 import FieldGroup from '../FieldGroup';
 import Button from '../../Button';
 import TextField from '../../TextField';
@@ -24,7 +24,7 @@ storiesOf('Components|Form', module)
       width: 400,
     }),
   )
-  .addDecorator(StateDecorator(store))
+  .addDecorator(StateDecorator(store) as StoryDecorator)
   .add(
     'default',
     withInfo()(() => (
@@ -60,14 +60,18 @@ storiesOf('Components|Form', module)
             labelText="I agree"
             value="yes"
             helpText="Click if you agree"
-            onChange={e => store.set({ agreeTerms: e.target.value })}
+            onChange={e =>
+              store.set({ agreeTerms: (e.target as HTMLInputElement).value })
+            }
             checked={store.state.agreeTerms === 'yes'}
             id="termsCheckboxYes"
           />
           <CheckboxField
             labelText="I don't agree"
             value="no"
-            onChange={e => store.set({ agreeTerms: e.target.value })}
+            onChange={e =>
+              store.set({ agreeTerms: (e.target as HTMLInputElement).value })
+            }
             checked={store.state.agreeTerms === 'no'}
             helpText="Click if you don't agree"
             id="termsCheckboxNo"
